Added Install Details
The previous README didn't have any information on getting the program up and running, and I ran into significant difficulties trying to install using the requirements.txt file. I later found out that it was because Anaconda by default uses the latest version of python, which was 3.13, and that version of python was so recent that it wasn't supported by even the most recent version of TensorFlow. I specifically say to download 3.10, but 3.11 might also work. I tested 3.9 and it didn't work and 3.12 also ran into troubles.

Also referenced the paper which used this code to ID the sea stars.